// utils/models/model.enums.ts
export enum SubscriptionStatus {
    FREE = "FREE",
    PREMIUM = "PREMIUM",
}

export enum SubscriptionFrequency {
    MONTHLY = "MONTHLY",
    YEARLY = "YEARLY",
}

export enum FriendStatus {
    PENDING = "PENDING",
    ACCEPTED = "ACCEPTED",
    REJECTED = "REJECTED",
}

export enum ConcentrationExercise {
    PODS = "PODS",
    PATTERNS = "PATTERNS",
    BREATHING = "BREATHING",
}

export enum Profession {
    LYCEEN = "LYCEEN",
    ETUDIANT = "ETUDIANT",
    ENTREPRENEUR = "ENTREPRENEUR",
    CHERCHEUR = "CHERCHEUR",
    MANAGER = "MANAGER",
    SALARIE = "SALARIE",
    AUTRE = "AUTRE",
}

export enum NotificationType {
    AUTO_SUGGESTIONS = "AUTO_SUGGESTIONS",
    MESSAGE = "MESSAGE",
    MAIL = "MAIL",
    FLASH = "FLASH",
}

export enum AgeRange {
    UNDER_18 = "UNDER_18",
    AGE_18_24 = "AGE_18_24",
    AGE_25_34 = "AGE_25_34",
    AGE_35_44 = "AGE_35_44",
    AGE_45_54 = "AGE_45_54",
    AGE_55_64 = "AGE_55_64",
    OVER_64 = "OVER_64",
}

export enum TimeSessionType {
    WORK = "WORK",
    BREAK = "BREAK",
}

export enum ExerciseFrequency {
    ONE_PER_SESSION = "ONE_PER_SESSION",
    AFTER_EACH_BREAK = "AFTER_EACH_BREAK",
}